# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "SLSEA Solar Generation API",
    "version": "1.0.0",
    "description": "RESTful API for Sri Lanka Sustainable Energy Authority solar generation telemetry, geographic hierarchy, and operational analytics."
  },
  "servers": [
    {
      "url": "https://slsea-solar-api-git-dev-diluka-theekshana-s-projects.vercel.app",
      "description": "Production Vercel Serverless"
    },
    {
      "url": "http://localhost:3000",
      "description": "Local Development Server"
    }
  ],
  "security": [
    {
      "UserAuth": []
    }
  ],
  "components": {
    "securitySchemes": {
      "UserAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-User-Id",
        "description": "User ID or Username for RBAC: 1 / national_admin, 2 / provincial_officer, 3 / district_operator"
      },
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key",
        "description": "API Key for meter telemetry device ingestion (e.g. key_inst_1)"
      }
    },
    "schemas": {
...
      "SolarInstallation": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "example": 1
          },
          "name": {
            "type": "string",
            "example": "Solar Installation 001"
          },
          "meter_id": {
            "type": "string",
            "example": "SLMTR-00001"
          },
          "inverter_id": {
            "type": "string",
            "example": "INV-00001"
          },
          "substation_id": {
            "type": "integer",
            "example": 1
          },
          "latitude": {
            "type": "number",
            "example": 6.939648
          },
          "longitude": {
            "type": "number",
            "example": 79.818451
          },
          "capacity_kw": {
            "type": "number",
            "example": 8
          }
        }
      },
...
      "CompositeInstallation": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "example": 1
          },
          "name": {
            "type": "string",
            "example": "Solar Installation 001"
          },
          "meter_id": {
            "type": "string",
            "example": "SLMTR-00001"
          },
          "inverter_id": {
            "type": "string",
            "example": "INV-00001"
          },
          "substation_id": {
            "type": "integer",
            "example": 1
          },
          "latitude": {
            "type": "number",
            "example": 6.939648
          },
          "longitude": {
            "type": "number",
            "example": 79.818451
          },
          "capacity_kw": {
            "type": "number",
            "example": 8
          },
          "last_reading": {
            "$ref": "#/components/schemas/GenerationReading",
            "nullable": true
          }
        }
      },
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "code": {
                "type": "string",
                "example": "RESOURCE_NOT_FOUND"
              },
              "message": {
                "type": "string",
                "example": "Province not found"
              }
            }
          }
        }
      },
...
    }
  },
  "paths": {
...
    "/installations": {
      "get": {
        "summary": "List all installations",
        "description": "Query all solar installations from MongoDB (api_key excluded)",
        "security": [
          {
            "UserAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Array of all solar installations",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SolarInstallation"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Solar Installations"
        ]
      }
    },
    "/installations/{id}": {
      "get": {
        "summary": "Get composite installation details",
        "description": "Composite resource embedding installation fields plus single latest reading",
        "security": [
          {
            "UserAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Composite installation object with last_reading",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CompositeInstallation"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Solar installation not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "tags": [
          "Solar Installations"
        ]
      },
      "put": {
        "summary": "Update solar installation",
        "description": "Update details of an existing solar installation",
        "security": [
          {
            "UserAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "meter_id": {
                    "type": "string"
                  },
                  "inverter_id": {
                    "type": "string"
                  },
                  "substation_id": {
                    "type": "integer"
                  },
                  "latitude": {
                    "type": "number"
                  },
                  "longitude": {
                    "type": "number"
                  },
                  "capacity_kw": {
                    "type": "number"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated installation object",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SolarInstallation"
                }
              }
            }
          },
          "400": {
            "description": "Validation failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Solar installation not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "tags": [
          "Solar Installations"
        ]
      },
      "delete": {
        "summary": "Soft delete solar installation",
        "description": "Soft delete a solar installation record",
        "security": [
          {
            "UserAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Installation soft deleted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Solar installation deleted successfully"
                    },
                    "id": {
                      "type": "integer",
                      "example": 1
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Solar installation not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "tags": [
          "Solar Installations"
        ]
      }
    },
    "/installations/{id}/readings/latest": {
      "get": {
        "summary": "Get latest generation reading",
        "description": "Derived operational real-time view returning the most recent reading",
        "security": [
          {
            "UserAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Bare latest generation reading object",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenerationReading"
                }
              }
            }
          },
          "404": {
            "description": "Solar installation or readings not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "tags": [
          "Telemetry Readings"
        ]
      }
    },
    "/installations/{id}/readings": {
      "get": {
        "summary": "Get analytical generation reading history",
        "description": "Historical readings with pagination envelope, date range filtering, and conditional GET ETag support",
        "security": [
          {
            "UserAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 1
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 50,
              "maximum": 100
            }
          },
          {
            "name": "from",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "format": "date-time"
            },
            "description": "Start ISO date-time filter"
          },
          {
            "name": "to",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "format": "date-time"
            },
            "description": "End ISO date-time filter"
          },
          {
            "name": "sort",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "timestamp"
            }
          },
          {
            "name": "order",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "asc",
                "desc"
              ],
              "default": "desc"
            }
          },
          {
            "name": "If-None-Match",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "ETag header for conditional GET"
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated envelope response",
            "headers": {
              "ETag": {
                "schema": {
                  "type": "string"
                }
              },
              "Cache-Control": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReadingsHistoryEnvelope"
                }
              }
            }
          },
          "304": {
            "description": "Not Modified (ETag matched)",
            "headers": {
              "ETag": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Solar installation not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "tags": [
          "Telemetry Readings"
        ]
      },
      "post": {
        "summary": "Ingest meter telemetry reading",
        "description": "Telemetry write endpoint requiring X-API-Key header authentication",
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateReadingPayload"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Reading created successfully",
            "headers": {
              "Location": {
                "schema": {
                  "type": "string"
                }
              },
              "ETag": {
                "schema": {
                  "type": "string"
                }
              },
              "Last-Modified": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenerationReading"
                }
              }
            }
          },
          "400": {
            "description": "Invalid reading payload fields",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "X-API-Key header is required for device ingestion",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Provided API key does not match this installation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Solar installation not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "tags": [
          "Telemetry Readings"
        ]
      }
    },
...
    "/readings/{id}": {
      "get": {
        "summary": "Get a single generation reading by ID",
        "description": "Fetch a specific telemetry reading by its unique numeric ID",
        "security": [
          {
            "UserAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Bare generation reading object",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenerationReading"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Generation reading not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "tags": [
          "Telemetry Readings"
        ]
      }
    },
...
  },
...
}